// MCP服务器/finance/shlomiuziel_asher-mcp_8cc730e_typescript/src/types.ts
export interface HapoalimCredentials {
  userCode: string;
  password: string;
}

export interface LeumiCredentials {
  username: string;
  password: string;
}

export interface DiscountCredentials {
  id: string;
  password: string;
  num: string;
}

export interface MercantileCredentials {
  id: string;
  password: string;
  num: string;
}

export interface MizrahiCredentials {
  username: string;
  password: string;
}

export interface BeinleumiCredentials {
  username: string;
  password: string;
}

export interface MassadCredentials {
  username: string;
  password: string;
}

export interface OtsarHahayalCredentials {
  username: string;
  password: string;
}

export interface VisaCalCredentials {
  username: string;
  password: string;
}

export interface MaxCredentials {
  username: string;
  password: string;
}

export interface IsracardCredentials {
  id: string;
  card6Digits: string;
  password: string;
}

export interface AmexCredentials {
  username: string;
  card6Digits: string;
  password: string;
}

export interface YahavCredentials {
  username: string;
  password: string;
  nationalID: string;
}

export interface BeyhadBishvilhaCredentials {
  id: string;
  password: string;
}